swig-templates/requirements.template.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
gevent==1.3.7
greenlet==0.4.15
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
locustio==0.13.5
MarkupSafe==1.1.0
msgpack==0.5.6
pyzmq==17.1.2
requests==2.20.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1